First attempt making pokemon as objects
experimental attempt at making pokemon as object in an OOP style.
The 3 files in the pokedex folder will contain all of the base models and
constants for creating a pokemon, then from there we can define all of the
pokemon as unique objects in the Gen 1 folder. If a pokemon is changed
from Gen 1 to Gen X you can use inheritence to deal with the changes.